+++ locals.tf
@@ -12,4 +12,25 @@ locals {
     ]
   ])
   subscriptions = { for sub in local.topics_subs : "${sub.topic_name}.${sub.sub_name}" => sub }
+
+  queues_auth = flatten([
+    for q_name, q in local.queues : [
+      for rule in ["listen", "send", "manage"] : {
+        queue          = q_name
+        rule           = rule
+        custom_name    = q.custom_name
+        authorizations = q.authorizations
+      }
+    ]
+  ])
+  topics_auth = flatten([
+    for t_name, t in local.topics : [
+      for rule in ["listen", "send", "manage"] : {
+        topic          = t_name
+        rule           = rule
+        custom_name    = t.custom_name
+        authorizations = t.authorizations
+      }
+    ]
+  ])
 }